cleaning - extract transaction to a separate class
